Replaced switch statements with if & else statements as Resource IDs will be non-final in Android Gradle Plugin version 5.0.

+++ app/src/main/java/io/github/waikato_ufdl/ui/images/ImageListAdapter.java
@@ -422,24 +422,15 @@ public class ImageListAdapter extends ListAdapter<ClassifiedImage, ImageListAdap
             }
 
             //check which menu item was clicked
-            switch (item.getItemId()) {
-                case R.id.action_delete:
-                    //when user presses delete
-                    deleteConfirmation(mode);
-                    break;
-
-                case R.id.action_relabel:
-                    //when the user presses edit
-                    confirmEditCategories(mode);
-                    break;
-
-                case R.id.action_select_all:
-                    selectAll();
-                    break;
-
-                case R.id.action_classify:
-                    showAutoClassifierDialog(mode);
-                    break;
+            int itemId = item.getItemId();
+            if (itemId == R.id.action_delete) {//when user presses delete
+                deleteConfirmation(mode);
+            } else if (itemId == R.id.action_relabel) {//when the user presses edit
+                confirmEditCategories(mode);
+            } else if (itemId == R.id.action_select_all) {
+                selectAll();
+            } else if (itemId == R.id.action_classify) {
+                showAutoClassifierDialog(mode);
             }
 
             return false;
